buspirate: Refactor how modes are modeled
The previous approach of putting all of the modes on a single struct and
using a type paramter to distinguish them ended up being quite confusing
in practice, because the modes all have distinct interfaces.

Instead, we'll use a separate named struct per mode so that it's clearer
which operations belong to which modes.